F one-infra#178: Remove ONEGATE_ENABLE input
The user input 'ONEGATE_ENABLE' was used as a crutch in the template to
setup both the 'TOKEN' and 'REPORT_READY' to 'YES' in a convenient way.

The issue is that if the template is not left in the default state then
the following content could be erased:

  REPORT_READY = "$ONEGATE_ENABLE",
  TOKEN = "$ONEGATE_ENABLE"

And 'ONEGATE_ENABLE' would be rendered defunct - which may confuse the
user: why the behavior of the 'ONEGATE_ENABLE' input does not work
anymore.

This can happen when 'Add OneGate Token' or 'Report Ready to OneGate' is
changed in the Sunstone and the original template workaround with
'ONEGATE_ENABLE' is broken:

  ONEGATE_ENABLE = "$ONEGATE_ENABLE",
  REPORT_READY = "YES",
  TOKEN = "YES"

The VM template is left with disabled OneGate by default (as is with the
all other appliances) but it is enabled be default for the Service
Template - where OneGate is needed for it to work properly in OneFlow
scenario.

Signed-off-by: Petr Ospalý <pospaly@opennebula.io>